flake: init new
The input `nixpkgs` is commented out because we don't yet need it for
anything. The `lib` comes from `birdos`, which should probably be
replaced in the future so that this flake doesn't pull my personal
dotfiles into hundreds of people's store. How to handle this is still in
question, but I think `birdos.lib` ought to be split out in a similar
fasion to this flake (the HM module comes from `spikespaz/dotfiles`).

The supported `systems` are hard-coded to `[ "x86_64-linux" ]`. This
needs to be corrected later on. We can use `nix-systems/default` after
making a pull request to `hyprwm/hyprland` to do the same.