flake: export overlays, packages, nixConfig
Exported from `hyprwm/hyprland` flake for now until we determine how to
handle the package. Re-exporting them from the original flake is
probably the best way to go, now and in-future.

The outputs `devShells` and `nixosModules` are missing because they are
outside of the scope of this project. The `devShells` should be kept
with the package, which is in `hyprwm/hyprland`. The `nixosModules` will
eventually be exported from this flake, and will share much common code
with `homeManagerModules`.