hm-mod: environment: fix compatibility with HM systemd options
This change introduces:
- `wayland.windowManager.hyprland.systemd.enable`
- `wayland.windowManager.hyprland.systemd.variables`
- `wayland.windowManager.hyprland.systemd.extraCommands`

This helps maintain parity with the current `hyprland` module in Home
Manager. I discovered that some other modules (namely `way-displays`)
expects `systemd.variables` to exist, and the lack of the option was
preventing my system from building.

The introduction of `systemd.enable` replaces `systemdIntegration`.
This was a holdover because Hyprnix never tracked [this
commit](https://github.com/nix-community/home-manager/commit/ed0770e96225f998ea128549ad446d9b1568cb2c).

The other two options introduced here, `systemd.variables` and
`systemd.extraCommands`, are additive, and don't replace other options.
Instead, there's new merging behavior (very simple) to keep
functionality in-line with `dbusEnviroment`/`extraDbusEnvironment`.

Since `dbus-update-activation-environment` can be used outside of a
systemd unit, it makes sense to keep the `dbusEnvironment` option for
compatibility with other init systems. Now, because of this, the
behavior is to make two calls to `dbus-update-activation-environment`,
the first one behaves normally, simply exporting `dbusEnvironment` and
`extraDbusEnvironment`, and the second call adds in `systemd.variables`
along with the `--systemd` flag. This is redundant, but that shouldn't
be a problem, because updates made by
`dbus-update-activation-environment` are additive and don't unset old
variables, so the effects of setting it twice are cumulative.

Do note that `systemd.variables` is empty by default and the
documentation suggests that one should instead use
`extraDbusEnvironment`. This makes it clear that the option is only
provided for compatibility with existing Home Manager modules.

Lastly, the introduction of `extraCommands` implies a behavioral change
compared to previous implementation, as before it would always run
`systemctl --user start hyprland-session.target`, but now it is
configurable, running `stop` and then `start` by default. I'm not sure
why the `stop` is necessary, but I deferred this judgement to the commit
previously mentioned.