hm-mod: monitors: init mkIntEnum factory
This factory produces the `type` and `apply` functions for `transform`
and `vrrMode` options.

The type of `transformEnum`'s value has changed, it is now an attribute
set of `mapping`, `type`, and `apply`, of which the latter two are
inherited in the `transform` option's `mkOption` arguments.

Also introduces `vrrModeEnum`, which has the same general form.

There is no longer a need for `transformEnumType`, so it has been
removed.

The moment this option type factory is used outside of `monitors.nix`,
it will move to `lib.hyprnix.types`. For now, it's not used anywhere
else.